First find the file, then open it
