+++ test/integration/007_open_find_test/local_dependency/dbt_project.yml
