+++ test/integration/007_open_find_test/local_dependency/models/my_package_model.sql
