+++ test/integration/007_open_find_test/models/my_model.sql
